requirements.txt
mock==2.0.0
requests==2.13.0
python-vlc==1.1.2